Move to utils.diff
